# src/swagger.json
{
	"components": {
...
		"schemas": {
			"IApiResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {},
					"count": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IErrorResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"details": {
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IVehicle": {
				"properties": {
					"make": {
						"type": "string"
					},
					"model": {
						"type": "string"
					},
					"chasis_number": {
						"type": "string"
					},
					"plate_number": {
						"type": "string"
					},
					"registration_date": {
						"type": "string"
					},
					"municipal_assembly": {
						"type": "string"
					},
					"terminal": {
						"type": "string"
					},
					"owner": {
						"type": "string"
					},
					"driver": {
						"type": "string"
					}
				},
				"required": [
					"make",
					"model",
					"chasis_number",
					"plate_number",
					"registration_date"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "app-engine",
		"version": "1.0.0",
		"description": "Taxi Check app engine to house all services (Monolith)",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "iCode"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/vehicle/all": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiResponse"
								}
							}
						}
					},
					"404": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Vehicle"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/vehicle/create": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created"
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Vehicle"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IVehicle"
							}
						}
					}
				}
			}
		},
		"/api/vehicle/getByNumberPlate/{numberPlate}": {
			"get": {
				"operationId": "GetByNumberPlate",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Vehicle"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "numberPlate",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/vehicle/getById/{vehicleId}": {
			"get": {
				"operationId": "FindVehicleById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IApiResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Vehicle"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "vehicleId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}